requirements01.txt
asgiref==3.5.2
boto3==1.24.76
botocore==1.27.76
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Django==3.2.13
django-cors-headers==3.13.0
django-crontab==0.7.1
django-filter==22.1
django-storages==1.13.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
install==1.3.5
jmespath==1.0.1
Markdown==3.3.7
numpy==1.23.3
Pillow==9.2.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
zipp==3.8.0
# =================
opencv-python-headless
numpy==1.23.3
face_recognition==1.2.3